Include transform changes in list that cause incremental reflows.
Also ensure that 3d translations get layers.

Fixes #8329.
