+++ components/layout/block.rs
@@ -55,7 +55,7 @@ use std::cmp::{max, min};
 use std::fmt;
 use std::sync::Arc;
 use style::computed_values::{border_collapse, box_sizing, display, float, overflow_x, overflow_y};
-use style::computed_values::{position, text_align, transform, transform_style};
+use style::computed_values::{position, text_align, transform_style};
 use style::properties::ComputedValues;
 use style::values::computed::{LengthOrNone, LengthOrPercentageOrNone};
 use style::values::computed::{LengthOrPercentage, LengthOrPercentageOrAuto};
@@ -583,33 +583,6 @@ impl BlockFlow {
                 BlockType::NonReplaced
             }
         }
-    }
-
-    pub fn transform_requires_layer(&self) -> bool {
-        // Check if the transform matrix is 2D or 3D
-        if let Some(ref transform_list) = self.fragment.style().get_effects().transform.0 {
-            for transform in transform_list {
-                match *transform {
-                    transform::ComputedOperation::Perspective(..) => {
-                        return true;
-                    }
-                    transform::ComputedOperation::Matrix(m) => {
-                        // See http://dev.w3.org/csswg/css-transforms/#2d-matrix
-                        if m.m31 != 0.0 || m.m32 != 0.0 ||
-                           m.m13 != 0.0 || m.m23 != 0.0 ||
-                           m.m43 != 0.0 || m.m14 != 0.0 ||
-                           m.m24 != 0.0 || m.m34 != 0.0 ||
-                           m.m33 != 1.0 || m.m44 != 1.0 {
-                            return true;
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-
-        // Neither perspective nor transform present
-        false
     }
 
     /// Compute the actual inline size and position for this block.
@@ -1560,7 +1533,7 @@ impl BlockFlow {
 
         // This flow needs a layer if it has a 3d transform, or provides perspective
         // to child layers. See http://dev.w3.org/csswg/css-transforms/#3d-rendering-contexts.
-        let has_3d_transform = self.transform_requires_layer();
+        let has_3d_transform = self.fragment.style().transform_requires_layer();
         let has_perspective = self.fragment.style().get_effects().perspective !=
             LengthOrNone::None;
